src/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
fastapi==0.121.1
idna==3.11
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
PuLP==3.3.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
